refactor: remove video highlight agent and integrate highlight agent for video highlights
- Deleted the `video_highlight_agent.py` file and its associated logic.
- Updated `router_collector.py` to use `HighlightAgent` instead of `VideoHighlightAgent`.
- Removed error handling for highlights in `page.tsx` and adjusted loading state management.
- Refactored `HighlightsCarousel` to handle video playback in a modal dialog, supporting YouTube and MP4 formats.
